Fixed LUT interpolation, improved performance, cleaned code,...

+++ shaders/bufferSettings.glsl
@@ -20,7 +20,7 @@ const int colortex2Format         = RGBA16F;
 
 SHADOWMAP & MISC
 const int shadowcolor0Format      = RGBA16F;
-const int colortex3Format         = RGBA32F;
+const int colortex3Format         = RGBA16F;
 
 MAIN COLOR
 const int colortex4Format         = RGBA32F;
@@ -31,13 +31,13 @@ const int colortex5Format         = RGBA32F;
 const bool colortex5Clear         = false;
 
 SKY ILLUMINANCE
-const int colortex6Format         = RGB32F;
+const int colortex6Format         = RGB16F;
 
 HISTORY
 const int colortex8Format         = RGBA32F;
 const bool colortex8Clear         = false;
 
-const int colortex9Format         = RGBA32F;
+const int colortex9Format         = RGBA16F;
 const bool colortex9Clear         = false;
 
 const int colortex10Format        = RGBA32F;
@@ -50,6 +50,6 @@ const int colortex12Format        = RGBA32F;
 const bool colortex12Clear        = false;
 
 CLOUDS
-const int colortex15Format        = RGBA32F;
+const int colortex15Format        = RGBA16F;
 const bool colortex15Clear        = false;
 */
